Rename SeeMoreButton to SeeMoreLink
The distinction between a button and a link is semantically important to
screen readers. Links load a new URL while buttons cause something to
happen on the current page. An example of a button might be a close
button, since closing a modal is a change to the current page. This
distinction should be explicitly maintained in the naming of these
components.

If a link component needs to act like a button or vice versa, aria-role
can be added to make the difference clear (WEB-2051).